feat: deixar que apenas o admin q postou possa apagar/mudar sua postagem

+++ app/Http/Controllers/ParagraphController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Paragraph;
 use App\Models\Photo;
+use App\Models\Post;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,6 +19,9 @@ class ParagraphController extends Controller
     public function index()
     {
         $paragraphs = Paragraph::with('photos')->get();
+        if (!$paragraphs) {
+            return response()->json(['messagem' => 'Paragrafos não encontrados'], Response::HTTP_NO_CONTENT);
+        }
         return response()->json(['data' => $paragraphs], Response::HTTP_OK);
     }
 
@@ -68,7 +72,6 @@ class ParagraphController extends Controller
         if (!$paragraph) {
             return response()->json(['messagem' => 'Parágrafo não encontrado'], Response::HTTP_NO_CONTENT);
         }
-
         return response()->json(['data' => $paragraph], Response::HTTP_OK);
     }
 
@@ -93,28 +96,32 @@ class ParagraphController extends Controller
             return response()->json(['Erro' => 'Impossível realizar a atualização, postagem não encontrada'], Response::HTTP_NO_CONTENT);
         }
 
-        $paragraph->update([
-            'subtitle' => $request->subtitle ?: $paragraph->subtitle,
-            'content' => $request->text ?: $paragraph->content
+        $post = Post::find($paragraph->post_id);
 
-        ]);
+        if ($post->user_id == $request->user()->id) {
+            $paragraph->update([
+                'subtitle' => $request->subtitle ?: $paragraph->subtitle,
+                'content' => $request->text ?: $paragraph->content
 
-        if ($request->hasFile('photo')) {
-            foreach ($paragraph->photos as $photo) {
-                Storage::disk('public')->delete($photo->photo);
-                $photo->delete();
+            ]);
+
+            if ($request->hasFile('photo')) {
+                foreach ($paragraph->photos as $photo) {
+                    Storage::disk('public')->delete($photo->photo);
+                    $photo->delete();
+                }
+                foreach ($request->file('photo') as $newPhoto) {
+                    $image_url = $newPhoto->store('photos', 'public');
+
+                    Photo::create([
+                        'paragraph_id' => $paragraph->id,
+                        'photo' => $image_url
+                    ]);
+                }
             }
-            foreach ($request->file('photo') as $newPhoto) {
-                $image_url = $newPhoto->store('photos', 'public');
-                
-                Photo::create([
-                    'paragraph_id' => $paragraph->id,
-                    'photo' => $image_url
-                ]);
-            }
+            return response()->json(['data' => $paragraph->load('photos')], Response::HTTP_OK);
         }
-
-        return response()->json(['data' => $paragraph->load('photos')], Response::HTTP_OK);
+        return response()->json(['mensagem' => 'Você não tem permissão para realizar esta ação']);
     }
 
     /**
@@ -123,19 +130,25 @@ class ParagraphController extends Controller
      * @param  \App\Models\Paragraph  $paragraph
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
         $paragraph = Paragraph::find($id);
         if (!$paragraph) {
             return response()->json(['Erro' => 'Paragrafo não encontrado'], Response::HTTP_NO_CONTENT);
         }
-        foreach($paragraph->photos as $photo){
-            Storage::disk('public')->delete($photo->photo);
-            $photo->delete();
+
+        $post = Post::find($paragraph->post_id);
+
+        if ($post->user_id == $request->user()->id) {
+            foreach ($paragraph->photos as $photo) {
+                Storage::disk('public')->delete($photo->photo);
+                $photo->delete();
+            }
+
+            $paragraph->delete();
+
+            return response()->json(['messagem' => 'Parágrafo deletado com sucesso']);
         }
-
-        $paragraph->delete();
-
-        return response()->json(['messagem' => 'Parágrafo deletado com sucesso']);
+        return response()->json(['mensagem' => 'Você não tem permissão para realizar esta ação'], Response::HTTP_NO_CONTENT);
     }
 }
